+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/ContributionProvidersBoxTestGenerated.java
@@ -265,6 +265,12 @@ public class ContributionProvidersBoxTestGenerated extends AbstractContributionP
     }
 
     @Test
+    @TestMetadata("defaultBinding.kt")
+    public void testDefaultBinding() {
+      runTest("compiler-tests/src/test/data/box/aggregation/contributionproviders/defaultBinding.kt");
+    }
+
+    @Test
     @TestMetadata("defaultParameters.kt")
     public void testDefaultParameters() {
       runTest("compiler-tests/src/test/data/box/aggregation/contributionproviders/defaultParameters.kt");