+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/ir.kt
@@ -10,6 +10,7 @@ import dev.zacsweers.metro.compiler.compat.CompatContext
 import dev.zacsweers.metro.compiler.computeMetroDefault
 import dev.zacsweers.metro.compiler.exitProcessing
 import dev.zacsweers.metro.compiler.expectAsOrNull
+import dev.zacsweers.metro.compiler.filterToSet
 import dev.zacsweers.metro.compiler.fir.MetroDiagnostics
 import dev.zacsweers.metro.compiler.fir.annotationsIn
 import dev.zacsweers.metro.compiler.fir.isExtensionGenerated
@@ -1788,7 +1789,7 @@ private fun List<IrConstructorCall>?.annotationsAnnotatedWith(
   annotationsToLookFor: Collection<ClassId>
 ): Set<IrConstructorCall> {
   if (this == null) return emptySet()
-  return filterTo(LinkedHashSet()) {
+  return filterToSet {
     it.type.classOrNull?.owner?.isAnnotatedWithAny(annotationsToLookFor) == true
   }
 }
@@ -2014,22 +2015,33 @@ internal fun Collection<IrClass>.toIrVararg() = ifNotEmpty {
 
 // Also check ignoreQualifier for interop after entering interop block to prevent unnecessary
 // checks for non-interop
-context(context: IrPluginContext)
-internal fun IrConstructorCall.bindingTypeOrNull(): Pair<IrType?, Boolean> {
-  return bindingTypeArgument()?.let { type ->
+context(context: IrMetroContext)
+internal fun IrConstructorCall.bindingTypeOrNull(
+  contributingClass: IrClass,
+  ignoreQualifier: Boolean,
+): IrTypeKey? {
+  bindingTypeArgument()?.let { typeKey ->
     // Return a binding defined using Metro's API
-    type to false
+    return typeKey.copy(qualifier = typeKey.qualifier ?: contributingClass.qualifierAnnotation())
   }
-    ?:
-    // Return a boundType defined using anvil KClass
-    (anvilKClassBoundTypeArgument() to anvilIgnoreQualifier())
+  val anvilBoundType = anvilKClassBoundTypeArgument() ?: return null
+  // Return a boundType defined using anvil KClass
+  val qualifier =
+    if (!ignoreQualifier) {
+      contributingClass.qualifierAnnotation()
+    } else {
+      null
+    }
+  return IrTypeKey(anvilBoundType, qualifier)
 }
 
-context(context: IrPluginContext)
-internal fun IrConstructorCall.bindingTypeArgument(): IrType? {
+context(context: IrMetroContext)
+internal fun IrConstructorCall.bindingTypeArgument(): IrTypeKey? {
   return getValueArgument(Symbols.Names.binding)?.expectAsOrNull<IrConstructorCall>()?.let {
     bindingType ->
-    bindingType.typeArguments.getOrNull(0)?.takeUnless { it == context.irBuiltIns.nothingType }
+    val type =
+      bindingType.typeArguments.getOrNull(0)?.takeUnless { it == context.irBuiltIns.nothingType }
+    type?.let { IrTypeKey(it, it.qualifierAnnotation()) }
   }
 }